+++ src/vilkaar/index.js
@@ -1,59 +1,5 @@
-/**
- * @see https://confluence.adeo.no/display/TEESSI/Kodeverk+i+Melosys#KodeverkiMelosys-Vilk%C3%A5rene
- * @type {*[]}
- */
-const vilkaar = [
-  {
-    kode: 'FO_883_2004_INNGANGSVILKAAR',
-    term: null,
-  },
-  {
-    kode: 'FO_883_2004_ART12_1',
-    term: null,
-  },
-  {
-    kode: 'FO_883_2004_ART12_2',
-    term: null,
-  },
-  {
-    kode: 'FO_883_2004_ART16_1',
-    term: null,
-  },
-  {
-    kode: 'ART12_1_VESENTLIG_VIRKSOMHET',
-    term: null,
-  },
-  {
-    kode: 'ART12_1_FORUTGAAENDE_MEDLEMSKAP',
-    term: null,
-  },
-  {
-    kode: 'ART12_2_NORMALT_DRIVER_VIRKSOMHET',
-    term: null,
-  },
-  {
-    kode: 'FO_883_2004_ART87A',
-    term: null,
-  },
-  {
-    kode: 'BOSATT_I_NORGE',
-    term: null,
-  },
-  {
-    kode: 'FTRL_2_12_UNNTAK_TURISTSKIP',
-    term: null,
-  },
-  {
-    kode: 'FO_883_2004_ART11_3A',
-    term: null,
-  },
-  {
-    kode: 'FO_883_2004_ART11_4_1',
-    term: null,
-  },
-  {
-    kode: 'FO_883_2004_ART11_4_2',
-    term: null,
-  }
-];
-module.exports.vilkaar = vilkaar;
+const { vilkaar } = require('./vilkaar');
+
+module.exports = {
+  vilkaar,
+};
